+++ R/connection.R
@@ -142,10 +142,14 @@ connection <- function(origin, destination, datetime = Sys.time(),
     routes <- .connection_summary(routes)
   }
 
+  # Bug of data.table and sf combination? Drops sfc class, when only one row...
+  routes <- as.data.frame(routes)
+  routes$geometry <- sf::st_sfc(routes$geometry, crs = 4326)
+
   # Create sf object
   return(
     sf::st_as_sf(
-      as.data.frame(routes),
+      routes,
       sf_column_name = "geometry",
       crs = 4326
     )