+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Use *styler* package and use `tyler::tidyverse_style()`to format the package.
 * All *lintr* issues are solved, except from line length issues (limit of 80 characters).
+* Avoid drop of the `sfc` class of the geometry column when there is only one geometry/row in a `data.table` (closes [#111](https://github.com/munterfinger/hereR/issues/111)).
 
 # version 0.6.0